+++ CHANGES.rst
@@ -13,6 +13,8 @@ New Features
 - Added ``block_replicate``, ``block_reduce`` and ``block_average`` functions.
   [#402]
 
+- Added ``median_filter`` function. [#415, #371]
+
 
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^